remove replay
